refactor: Create step_1 subcommand
