+++ packages/cts1_mo_tools/pyproject.toml
@@ -28,7 +28,7 @@ cts1_agenda_maker = "cts1_mo_tools.cts1_agenda_maker.main:main"
 cts1_make_bulk_uplink_agenda = "cts1_mo_tools.cts1_make_bulk_uplink_agenda:main"
 cts1_spreadsheet_to_agenda = "cts1_mo_tools.cts1_spreadsheet_to_agenda:main"
 cts1_satnogs_interval = "cts1_mo_tools.cts1_satnogs_interval:main"
-cts1_processing_pipeline = "cts1_mo_tools.cts1_processing_pipeline.pipeline:main"
+cts1_processing_pipeline = "cts1_mo_tools.cts1_processing_pipeline.cli:main"
 
 [dependency-groups]
 dev = [
